Script to force re-sync all topics

+++ scripts/acl_cleaner.py
@@ -1,4 +1,5 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
+import argparse
 import os
 from dataclasses import dataclass
 from fnmatch import fnmatch
@@ -35,11 +36,12 @@ class Service:
 class AivenKafka(object):
     base = "https://api.aiven.io/v1/project"
 
-    def __init__(self, project, service=None):
+    def __init__(self, project, service=None, dry_run=False):
         self.project = project
         if service is None:
             service = project + "-kafka"
         self.service = service
+        self.dry_run = dry_run
         self.session = requests.Session()
         self.session.auth = AivenAuth()
         self.base_url = f"{self.base}/{self.project}/service/{self.service}"
@@ -67,20 +69,28 @@ class AivenKafka(object):
 
     def delete_acls(self, acls_to_delete):
         for acl in acls_to_delete:
-            print(f"Deleting {acl}")
-            resp = self.session.delete(f"{self.base_url}/acl/{acl.id}")
-            resp.raise_for_status()
+            if not self.dry_run:
+                print(f"Deleting {acl}")
+                resp = self.session.delete(f"{self.base_url}/acl/{acl.id}")
+                resp.raise_for_status()
+            else:
+                print(f"Would have deleted {acl}")
 
     def delete_users(self, users_to_delete):
         for username in users_to_delete:
-            print(f"Deleting {username}")
-            resp = self.session.delete(f"{self.base_url}/user/{username}")
-            resp.raise_for_status()
+            if not self.dry_run:
+                print(f"Deleting {username}")
+                resp = self.session.delete(f"{self.base_url}/user/{username}")
+                resp.raise_for_status()
+            else:
+                print(f"Would have deleted {username}")
 
 
 def find_unused_acls(topics, acls):
     for acl in acls:
-        if acl.topic.startswith("__") or acl.topic in topics:
+        if any((acl.topic.startswith("__"),
+                "_stream_" in acl.topic,
+                acl.topic in topics)):
             continue
         yield acl
 
@@ -94,8 +104,8 @@ def find_unused_users(acls, users):
             yield username
 
 
-def main():
-    aiven = AivenKafka("nav-prod")
+def main(project, dry_run):
+    aiven = AivenKafka(project, dry_run=dry_run)
 
     service = aiven.get_service()
 
@@ -110,4 +120,8 @@ def main():
 
 
 if __name__ == '__main__':
-    main()
+    parser = argparse.ArgumentParser()
+    parser.add_argument("-n", "--dry-run", action="store_true", help="Make no actual changes")
+    parser.add_argument("project", action="store", help="Aiven project to process")
+    options = parser.parse_args()
+    main(options.project, options.dry_run)
